Fixed the vector iterator bug!
Thank the lord for Matt Hokinson. Saved my butt hardcore. For some reason the ExerciseDatabase object wasn't building correctly in the app, even though I had it pretty much the same as what it was in the test. Changed the ExerciseDatabase to a local variable instead and it worked out.